fourth_commit

+++ src/main/java/hr/fer/ruazosa/trackmyroute/model/Route.java
@@ -30,7 +30,7 @@ public class Route {
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private Date date;
     @Column(name = "duration")
-    private Long duration;
+    private double duration;
     @Column(name = "speed")
     private double speed;
     @Column(name = "distance")
@@ -69,7 +69,7 @@ public class Route {
         this.date = date;
     }
 
-    public Long getDuration() {
+    public double getDuration() {
         return duration;
     }
